// src/gen/models/SHACLShape.ts
/* tslint:disable */
/* eslint-disable */
/**
 * AI assistant API model
 * This is the API for the AI assistant module
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: info@obdasystems.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface SHACLShape
 */
export interface SHACLShape {
    /**
     * 
     * @type {string}
     * @memberof SHACLShape
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof SHACLShape
     */
    type: SHACLShapeTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof SHACLShape
     */
    targetClass: string;
    /**
     * 
     * @type {string}
     * @memberof SHACLShape
     */
    path: string;
    /**
     * 
     * @type {string}
     * @memberof SHACLShape
     */
    property?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof SHACLShape
     */
    constraintValue?: Array<string>;
}

/**
* @export
* @enum {string}
*/
export enum SHACLShapeTypeEnum {
    CLASS = 'Class',
    MIN_COUNT = 'MinCount',
    MAX_COUNT = 'MaxCount',
    MIN_EXCLUSIVE = 'MinExclusive',
    MAX_EXCLUSIVE = 'MaxExclusive',
    MIN_INCLUSIVE = 'MinInclusive',
    MAX_INCLUSIVE = 'MaxInclusive',
    MIN_LENGTH = 'MinLength',
    MAX_LENGTH = 'MaxLength',
    PATTERN = 'Pattern',
    IN = 'In',
    EQUALS = 'Equals',
    DISJOINT = 'Disjoint',
    LESS_THAN = 'LessThan',
    LESS_THAN_OR_EQUALS = 'LessThanOrEquals'
}


/**
 * Check if a given object implements the SHACLShape interface.
 */
export function instanceOfSHACLShape(value: object): value is SHACLShape {
    if (!('type' in value) || value['type'] === undefined) return false;
    if (!('targetClass' in value) || value['targetClass'] === undefined) return false;
    if (!('path' in value) || value['path'] === undefined) return false;
    return true;
}

export function SHACLShapeFromJSON(json: any): SHACLShape {
    return SHACLShapeFromJSONTyped(json, false);
}

export function SHACLShapeFromJSONTyped(json: any, ignoreDiscriminator: boolean): SHACLShape {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'type': json['type'],
        'targetClass': json['targetClass'],
        'path': json['path'],
        'property': json['property'] == null ? undefined : json['property'],
        'constraintValue': json['constraintValue'] == null ? undefined : json['constraintValue'],
    };
}

export function SHACLShapeToJSON(json: any): SHACLShape {
    return SHACLShapeToJSONTyped(json, false);
}

export function SHACLShapeToJSONTyped(value?: SHACLShape | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'type': value['type'],
        'targetClass': value['targetClass'],
        'path': value['path'],
        'property': value['property'],
        'constraintValue': value['constraintValue'],
    };
}